// src/app/api/model/langue.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Livre } from './livre';


export interface Langue { 
    id?: number;
    codeIso?: string;
    livres?: Array<Livre>;
    nom?: string;
}